Add DPDT switch, reroute to account for switch
The switch swaps the A button and Start button, which makes it the same
as an Avenue Pad 3, basically. With the Avenue Pad 3, the extra button
exists where the A button is on the Genesis, and the mapping is selected
by a switch on the controller. Some games use III mapped to Sel, some
use III mapped to Run.

The original design had A (which you can consider III) hard-mapped to
Sel. Now it is selectable, which increases compatibility with games that
mapped III to Run instead.